+++ crates/board/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Major
 
+- Add `platform::Api::Update` (fix #47)
 - Add `Api::Platform` with `platform::Api::reboot()` function
 - Require the APIs to be `Send` and the top-level API to be `'static` too